BUGFIX: preferences API (self.package.{getPreference,setPreference}) didn't work.
NOTE: self.package.{getPreference,setPreference} is the only supported public API,
methods in the 'adm.preference' table (and the very existence and structure of
the table itself) should be considered an implementation detail.

CHANGED: sqliteadapter do not support multiple connections. TempEnv doesn't work.
Since it's used exclusively for testing (at the moment), TempEnv now is a no-op
with sqlite.

git-svn-id: http://svn.genropy.org/genro/trunk@3340 c8d40aac-abfe-4c07-9ed8-ca0f8f1311d8