added callAsync in gnrlang
callAsync(call=None, call_args=None, call_kwargs=None, cb=None, cb_args=None, cb_kwargs=None, exit_timeout = 60):
call is the method to call in a separate thread
call_args and call_kwargs are the args tuple and the kwargs dict for call
cb is a callback, with cb_args and cb_kwargs as args and kwargs.
cb_kwargs contains 'result' which is the return value of call.

Signed-off-by: Michele Bertoldi <michele.bertoldi@softwell.it>